+++ swov1/RELEASES.md
@@ -55,3 +55,13 @@ Based on:
 - [typescript v0.6.0] swov1
 ### Releases
 - [NPM v0.6.0] https://www.npmjs.com/package/@solarwinds/swo-sdk-typescript/v/0.6.0 - swov1
+
+## 2025-11-18 00:03:09
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.658.2 (2.755.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.6.1] swov1
+### Releases
+- [NPM v0.6.1] https://www.npmjs.com/package/@solarwinds/swo-sdk-typescript/v/0.6.1 - swov1